modified file

+++ prep_encap/Fraction.cs
@@ -1,25 +1,29 @@
 public class Fraction{
     // Attributes
-    private double numerator;
-    private double denominator;
+    private int numerator;
+    private int denominator;
 
     // Methods
     // Getter to return numerator value, if any
-    public double get_numerator(){
+    public int get_numerator(){
         return numerator;
     }
 
-    public double get_denominator(){
+    public int get_denominator(){
         return denominator;
     }
 
-    public double set_numerator(double numerator_set){
+    public int set_numerator(int numerator_set){
         numerator = numerator_set;
         return numerator;
     }
 
-    public double set_denominator(double denominator_set){
+    public int set_denominator(int denominator_set){
         denominator = denominator_set;
         return denominator;
     }
+
+    // public string get_fraction_string(int first_num, int second_num){
+
+    // }
 }
